Program.cs 역할별 디렉터리 분리
