+++ src/server-minimalapi/api/LocalLLMServer/Model/Dto/ChatRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace LocalLLMServer.Endpoint.Dto;
+
+public class ChatRequest
+{
+    public string Message { get; set; } = "";
+    public string Model { get; set; } = "";
+}
