+++ src/server-minimalapi/api/LocalLLMServer/Model/Dto/OllamaResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace LocalLLMServer.Endpoint.Dto;
+
+public class OllamaResponse
+{
+    public string response { get; set; } = "";
+}
