+++ src/server-minimalapi/api/LocalLLMServer/Model/Entity/FileName.cs
@@ -1,0 +1,9 @@
+﻿namespace LocalLLMServer.Model.Entity;
+
+public class OllamaModel
+{
+    public string Name { get; set; } = "";
+    public string Size { get; set; } = "";
+    public string LastUsed { get; set; } = "";
+    public string Status { get; set; } = "";
+}
